requirements.txt
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
requests==2.27.1
requests-oauthlib==1.3.1
tqdm==4.64.0